+++ src/experimental/event/test/interactive/CrosshairCanvas.js
