+++ src/experimental/event/test/interactive/prepareTestArea.js
