+++ src/experimental/event/test/interactive/testPages/newMouseOverMoveOut.js
@@ -1,10 +1,10 @@
 /* global document */
 
-import {addEvents} from '../../../src/core/event';
+import addEvents from '../../../add-events';
 import CrosshairCanvas from '../CrosshairCanvas';
 
 export default {
-  name: '(new) mouse enter/move/leave',
+  name: '(new) mouse over/move/out',
   render(testArea) {
     const containerElement = document.createElement('div');
     const canvasElement = document.createElement('canvas');
@@ -25,11 +25,11 @@ export default {
     `);
     const crosshairCanvas = new CrosshairCanvas(canvasElement);
     addEvents(canvasElement, {
-      onMouseEnter(eventInfo) {
+      onMouseOver(eventInfo) {
         console.log(eventInfo);
         crosshairCanvas.reset();
         crosshairCanvas.setCrosshair({
-          key: 'enter', position: eventInfo.pointerPosition, color: 'blue'
+          key: 'over', position: eventInfo.pointerPosition, color: 'blue'
         });
       },
       onMouseMove(eventInfo) {
@@ -38,10 +38,10 @@ export default {
           key: 'move', position: eventInfo.pointerPosition, color: 'black'
         });
       },
-      onMouseLeave(eventInfo) {
+      onMouseOut(eventInfo) {
         console.log(eventInfo);
         crosshairCanvas.setCrosshair({
-          key: 'leave', position: eventInfo.pointerPosition, color: 'red'
+          key: 'out', position: eventInfo.pointerPosition, color: 'red'
         });
       }
     });
