+++ src/experimental/index.js
@@ -1,1 +1,2 @@
 export {default as ShaderCache} from './shader-cache';
+export * from './event';
